requirements.txt
PyYAML==6.0.2
cx_Oracle==8.3.0
pyodbc==4.0.32
psycopg2==2.9.10
jinja2==3.1.4
FPE==0.1.2
asyncio==3.4.3
#logging==0.5.1.2
black==24.10.0
ff3
sphinx
